[Feat] MainViewController Network 관련 코드 ScheduleViewController로 이동
