+++ SobokSobok/SobokSobok/Presentation/Main/ViewControllers/MainViewController.swift
@@ -15,20 +15,6 @@ final class MainViewController: BaseViewController {
     private let scheduleViewController = ScheduleViewController()
     private let containerView = UIView()
     
-    let scheduleManager: ScheduleServiceable = ScheduleManager(apiService: APIManager(),
-                                                               environment: .mock)
-    var schedules: [Schedule] = [] {
-        didSet {
-            parseSchedules()
-        }
-    }
-    
-    var pillLists: [PillList] = [] {
-        didSet {
-            scheduleViewController.pillItems = pillLists
-        }
-    }
-    
     override func viewDidLoad() {
         super.viewDidLoad()
     }
@@ -36,11 +22,6 @@ final class MainViewController: BaseViewController {
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
         tabBarController?.tabBar.isHidden = false
-    }
-    
-    override func viewDidAppear(_ animated: Bool) {
-        getMySchedules(date: "2022-06-30")
-        getMyPillLists(date: "2022-06-22")
     }
 
     override func style() {
@@ -65,22 +46,5 @@ final class MainViewController: BaseViewController {
             $0.top.equalTo(homeTopView.snp.bottom)
             $0.leading.bottom.trailing.equalToSuperview()
         }
-    }
-}
-
-extension MainViewController {
-    func parseSchedules() {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = Date.FormatType.full.description
-
-        let doneItems = schedules
-            .filter { $0.isComplete == "done" }
-            .map { dateFormatter.date(from: $0.scheduleDate)!.toString(of: .year) }
-        let doingItems = schedules
-            .filter { $0.isComplete == "doing" }
-            .map { dateFormatter.date(from: $0.scheduleDate)!.toString(of: .year) }
-        
-        scheduleViewController.doneDates = doneItems
-        scheduleViewController.doingDates = doingItems
     }
 }
