+++ SobokSobok/SobokSobok/Presentation/Schedule/Controllers/ScheduleViewController+DataSource.swift
@@ -11,11 +11,11 @@ import UIKit
 
 extension ScheduleViewController: UICollectionViewDataSource {
     func numberOfSections(in collectionView: UICollectionView) -> Int {
-        return pillItems.count
+        return pillLists.count
     }
     
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return pillItems[section].scheduleList?.count ?? 0
+        return pillLists[section].scheduleList?.count ?? 0
     }
     
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
@@ -24,7 +24,7 @@ extension ScheduleViewController: UICollectionViewDataSource {
             for: indexPath
         ) as? ScheduleCell else { return UICollectionViewCell() }
         
-        if let pill = pillItems[indexPath.section].scheduleList?[indexPath.row] {
+        if let pill = pillLists[indexPath.section].scheduleList?[indexPath.row] {
             cell.configureCell(with: pill)
         }
